authorization using gate facade

+++ app/Http/Controllers/ImageController.php
@@ -33,17 +33,23 @@ class ImageController extends Controller
 
     public function edit(Image $image)
     {
+        $this->authorize('update-image', $image);
+
         return view("image.edit", compact('image'));
     }
 
     public function update(Image $image, ImageRequest $request)
     {
+        $this->authorize('update-image', $image);
+
         $image->update($request->getData());
         return to_route('images.index')->with('message', "Image has been updated successfully");
     }
 
     public function destroy(Image $image)
     {
+        $this->authorize('delete-image', $image);
+
         $image->delete();
         return to_route('images.index')->with('message', "Image has been removed successfully");
     }
